+++ src/api/models/__init__.py
@@ -1,9 +1,11 @@
 from api.models.article import Article
+from api.models.author import Author
 from api.models.blog import Blog
 from api.models.event import Event
 from api.models.launch import Launch
 from api.models.news_site import NewsSite
 from api.models.provider import Provider
 from api.models.report import Report
+from api.models.socials import Socials
 
-__all__ = ["Article", "Blog", "Event", "Launch", "NewsSite", "Provider", "Report"]
+__all__ = ["Article", "Author", "Blog", "Event", "Launch", "NewsSite", "Socials", "Provider", "Report"]