+++ src/api/serializers/author_serializer.py
@@ -1,0 +1,12 @@
+from rest_framework import serializers
+
+from api.models.author import Author
+from api.serializers.socials_serializer import SocialsSerializer
+
+
+class AuthorSerializer(serializers.ModelSerializer[Author]):
+    socials = SocialsSerializer()
+
+    class Meta:
+        model = Author
+        fields = ["name", "socials"]